+++ php/oop/medium/src/Comment.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App;
+
+use DateTimeImmutable;
+
+class Comment
+{
+    public function __construct(
+        private string $user,
+        private string $content,
+        private DateTimeImmutable $submitedAt
+    ) {
+    }
+}
