+++ php/oop/medium/src/Post.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App;
+
+class Post
+{
+    public function __construct(
+        private string $title,
+        private string $content,
+        private string $author,
+        private Comment $comment
+    ) {
+    }
+
+}
